+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for Ollama streaming with schema `OllamaSchema` (see `?StreamCallback` for more information). Schema `OllamaManaged` is NOT supported (it's legacy and will be removed in the future).
 - Moved the implementation of streaming callbacks to a new `StreamCallbacks` package.
 
+### Fixed
+- Fixed a bug in `call_cost` where the cost was not calculated if any non-AIMessages were provided in the conversation.
+
 ## [0.59.1]
 
 ### Fixed